services/sudar-tools-mcp-server/pyproject.toml
[project]
name = "sudar-tools-mcp-server"
version = "0.1.0"
description = "MCP server providing web search and scraping tools"
requires-python = ">=3.10"

dependencies = [
    "fastmcp>=0.2.0",
    "tavily-python>=0.5.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.32.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "minio>=7.2.0",
    "fastapi>=0.118.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
